Safari-611.1.10/LayoutTests/inspector/worker/dom-debugger-event-listener-breakpoints-expected.txt

Tests for listener event breakpoint functionality (DOMDebugger) in a Worker.


== Running test suite: Worker.DOMDebugger.EventBreakpoint.Listener
-- Running test case: Worker.DOMDebugger.EventBreakpoint.Listener.Basic
PASS: Should have paused.
PASS: Pause reason should have been Listener.
PASS: Pause data eventName should have been "custom".
CALL STACK:
0: [F] handler
1: [F] triggerListener
2: [F] handleMessage
ASYNC CALL STACK:
3: --- addEventListener ---
4: [P] Global Code
